+++ src/esm/drawutilssvg.js
@@ -115,7 +115,7 @@ export class drawutilssvg {
         // Which default styles to add? -> All from the DrawConfig.
         // Compare with DrawConfig interface
         const keys = {
-            // "bezier": "CubicBezierCurve", // TODO: is this correct?
+            "bezier": "CubicBezierCurve",
             "bezierPath": "BezierPath",
             "polygon": "Polygon",
             "triangle": "Triangle",